# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EstimateRequest": {
        "properties": {
          "exam_type": {
            "default": "Khám Bảo hiểm Y tế (BHYT)",
            "description": "Hình thức khám",
            "title": "Exam Type",
            "type": "string"
          },
          "is_bhyt": {
            "default": true,
            "description": "Có sử dụng thẻ BHYT hay không",
            "title": "Is Bhyt",
            "type": "boolean"
          },
          "services": {
            "description": "Danh sách tên các xét nghiệm/chụp chiếu cần làm",
            "items": {
              "type": "string"
            },
            "title": "Services",
            "type": "array"
          }
        },
        "title": "EstimateRequest",
        "type": "object"
      },
      "EstimateResponse": {
        "properties": {
          "examination_fee": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Examination Fee",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "services": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Services",
            "type": "array"
          },
          "total_estimated": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Estimated",
            "type": "string"
          }
        },
        "required": [
          "examination_fee",
          "services",
          "total_estimated",
          "note"
        ],
        "title": "EstimateResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepResponse"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "code",
          "name",
          "location",
          "steps"
        ],
        "title": "WorkflowResponse",
        "type": "object"
      },
      "WorkflowStepResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "step_number": {
            "title": "Step Number",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "step_number",
          "title",
          "detail",
          "location"
        ],
        "title": "WorkflowStepResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Healthcare Multi-Agent Chatbot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/hospital/estimate": {
      "post": {
        "description": "Ước tính chi phí khám và cận lâm sàng theo bảng giá Bệnh viện E.",
        "operationId": "estimate_cost_api_v1_hospital_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Estimate Cost",
        "tags": [
          "Bệnh viện E Tri thức"
        ]
      }
    },
    "/api/v1/hospital/workflows/{code}": {
      "get": {
        "description": "Lấy chi tiết quy trình khám bệnh (bhyt, on_demand, emergency).",
        "operationId": "get_workflow_api_v1_hospital_workflows__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow",
        "tags": [
          "Bệnh viện E Tri thức"
        ]
      }
    },
...
  }
}